+++ packages/api-client/src/index.ts
@@ -1,6 +1,7 @@
-export { createApiClient, getApiClientConfig } from './client-config';
+export { createApiClient, getApiClientConfig, setApiClientTenantId } from './client-config';
 export type { ApiClientConfig } from './client-config';
 export { customFetch } from './custom-fetch';
+export type { ApiFetchError } from './custom-fetch';
 
 export * from './generated/health/health';
 export * from './generated/tenants/tenants';